Implement dark theme and update language setting
This commit adds a dark theme to the application by creating a new theme file with appropriate color palettes. It also integrates this dark theme into the main layout file through the use of a ThemeProvider. Additionally, the language attribute in the HTML tag has been changed from English ('en') to German ('de').