Add form components and react-hook-form support
Implemented Form components and utilized the 'react-hook-form' library to develop a calculation form in the repayment calculator. Also created a schema for form validation and added appropriate utilities and select component for various form fields. The use of '@hookform/resolvers' is to facilitate 'zod' validation schema usage with the form.